refactor(store): move repair-shipping form components under components/
- presentational tsx 6개를 repair-shipping/components/로 이동 (커버리지 제외 glob src/**/components/**/*.tsx 컨벤션 복원)
- 로직 파일(pickup-request-model, upload-repair-shipping-photos, use-copy-repair-address)은 feature 루트 유지
- features/order/index.ts re-export 및 내부 상대 import 경로 수정

Confidence: high
Scope-risk: narrow
Tested: pnpm --dir apps/store tsc --noEmit; pnpm depcruise
Not-tested: 없음 (순수 파일 이동)
