+++ apps/store/src/features/order/index.ts
@@ -3,14 +3,14 @@ export { ReformOrderItemCard } from "./order-form/components/reform-order-item-c
 export {
   RepairAddressRows,
   RepairAddressCopyButton,
-} from "./repair-shipping/repair-address-rows";
+} from "./repair-shipping/components/repair-address-rows";
 export {
   TrackingModeToggle,
   type TrackingMode,
-} from "./repair-shipping/tracking-mode-toggle";
-export { TrackingFormFields } from "./repair-shipping/tracking-form-fields";
-export { NoTrackingFormFields } from "./repair-shipping/no-tracking-form-fields";
-export { PickupRequestFields } from "./repair-shipping/pickup-request-fields";
+} from "./repair-shipping/components/tracking-mode-toggle";
+export { TrackingFormFields } from "./repair-shipping/components/tracking-form-fields";
+export { NoTrackingFormFields } from "./repair-shipping/components/no-tracking-form-fields";
+export { PickupRequestFields } from "./repair-shipping/components/pickup-request-fields";
 export {
   EMPTY_PICKUP_REQUEST,
   type PickupRequestInfo,
@@ -18,7 +18,7 @@ export {
 export {
   RepairShippingMethodChoice,
   type RepairShippingMethod,
-} from "./repair-shipping/shipping-method-choice";
+} from "./repair-shipping/components/shipping-method-choice";
 export { uploadRepairShippingPhotos } from "./repair-shipping/upload-repair-shipping-photos";
 export { TokenRefundAction } from "./components/token-refund-action";
 export { isCustomOrderPaymentState } from "./custom-payment/types";
